+++ src/main/java/com/starkinfra/PixRequest.java
@@ -51,6 +51,7 @@ public final class PixRequest extends Resource {
      * tags [list of strings, default null]: list of strings for reference when searching for PixRequests. ex: ["employees", "monthly"]
      * method [string]: execution method for the creation of the Pix. ex: "manual", "payerQrcode", "dynamicQrcode".
      * priority [string, default "high"]: Specifies the message channel used to send the Pix Request. If set to "high", the request is sent through the primary channel; if set to "low", it uses the secondary channel. Options: "high" or "low"
+     * reason [string, default "customerRequest"]: underlying reason for the payment transaction. ex: "customerRequest", "fraud", "subscriptionFlaw"
      * id [string]: unique id returned when the PixRequest is created. ex: "5656565656565656"
      * fee [integer]: fee charged when PixRequest is paid. ex: 200 (= R$ 2.00)
      * status [string]: current PixRequest status. ex: "registered" or "paid"
@@ -86,6 +87,7 @@ public final class PixRequest extends Resource {
     public String[] tags;
     public String method;
     public String priority;
+    public String reason;
     public Integer fee;
     public String status;
     public String flow;
@@ -128,6 +130,7 @@ public final class PixRequest extends Resource {
      * @param tags [list of strings, default null]: list of strings for reference when searching for PixRequests. ex: ["employees", "monthly"]
      * @param method [string, default null]: execution method for the creation of the Pix. ex: "manual", "payerQrcode", "dynamicQrcode".
      * @param priority [string, default "high"]: Specifies the message channel used to send the Pix Request. If set to "high", the request is sent through the primary channel; if set to "low", it uses the secondary channel. Options: "high" or "low"
+     * @param reason [string, default "customerRequest"]: underlying reason for the payment transaction. ex: "customerRequest", "fraud", "subscriptionFlaw"
      * @param id [string]: unique id returned when the PixRequest is created. ex: "5656565656565656"
      * @param fee [integer]: fee charged when PixRequest is paid. ex: 200 (= R$ 2.00)
      * @param status [string]: current PixRequest status. ex: "registered" or "paid"
@@ -141,7 +144,7 @@ public final class PixRequest extends Resource {
                       String receiverBankCode, String receiverAccountNumber, String receiverBranchCode,
                       String receiverAccountType, String endToEndId, String receiverKeyId, String description,
                       String reconciliationId, String initiatorTaxId, Long cashAmount, String cashierBankCode,
-                      String cashierType, String[] tags, String method, String priority, Integer fee, String status, String flow,
+                      String cashierType, String[] tags, String method, String priority, String reason, Integer fee, String status, String flow,
                       String senderBankCode, String created, String updated, String id
     ) {
         super(id);
@@ -169,6 +172,7 @@ public final class PixRequest extends Resource {
         this.tags = tags;
         this.method = method;
         this.priority = priority;
+        this.reason = reason;
         this.fee = fee;
         this.status = status;
         this.flow = flow;
@@ -215,6 +219,7 @@ public final class PixRequest extends Resource {
      * tags [list of strings, default null]: list of strings for reference when searching for PixRequests. ex: ["employees", "monthly"]
      * method [string, default null]: execution method for the creation of the Pix. ex: "manual", "payerQrcode", "dynamicQrcode".
      * priority [string, default "high"]: Specifies the message channel used to send the Pix Request. If set to "high", the request is sent through the primary channel; if set to "low", it uses the secondary channel. Options: "high" or "low"
+     * reason [string, default "customerRequest"]: underlying reason for the payment transaction. ex: "customerRequest", "fraud", "subscriptionFlaw"
      * <p>
      * Attributes (return-only):
      * id [string]: unique id returned when the PixRequest is created. ex: "5656565656565656"
@@ -254,6 +259,7 @@ public final class PixRequest extends Resource {
         this.tags = (String[]) dataCopy.remove("tags");
         this.method = (String) dataCopy.remove("method");
         this.priority = (String) dataCopy.remove("priority");
+        this.reason = (String) dataCopy.remove("reason");
         this.fee = null;
         this.status = null;
         this.flow = null;